dos: add RenameRelative
