+++ rom/dos/dos.conf
@@ -267,4 +267,5 @@ BOOL SetFileDateRelative(BPTR lock, CONST_STRPTR name, const struct DateStamp *d
 BOOL SetOwnerRelative(BPTR lock, CONST_STRPTR name, ULONG owner_info) (D1,D2,D3)
 LONG SetProtectionRelative(BPTR lock, CONST_STRPTR name, ULONG protect) (D1,D2,D3)
 LONG MakeLinkRelative(BPTR lock, CONST_STRPTR name, SIPTR dest, LONG soft) (D1,D2,D3,D4)
+LONG RenameRelative(BPTR oldLock, CONST_STRPTR oldName, BPTR newLock, CONST_STRPTR newName) (D1,D2,D3,D4)
 ##end functionlist
