+++ rom/dos/mmakefile.src
@@ -56,7 +56,8 @@ FUNCTIONS := abortpkt addbuffers adddosentry addpart addsegment \
 	     matchpatternnocase maxcli namefromfh namefromlock newloadseg nextdosentry nil \
 	     open openfromlock openrelative output parentdir parentoffh parsepattern \
 	     parsepatternnocase pathpart printfault putstr read readargs \
-	     readitem relabel readlink remassignlist remdosentry remsegment rename \
+	     readitem relabel readlink remassignlist remdosentry remsegment \
+	     rename renamerelative \
 	     replypkt runcommand samedevice samelock scanvars seek \
 	     selectinput selectoutput sendpkt setargstr setcomment setcommentrelative \
 	     setconsoletask setcurrentdirname setfiledate setfiledaterelative \
